feat(ResultViewer): TreeViewer design changes - viewing mode
