+++ src/common/styles/variables.less
@@ -1,3 +1,5 @@
+// Color definitions
+
 @primary-tint-1: #dce5f5;
 @primary-tint-2: #bcccec;
 @primary-tint-3: #8a9dc7;
@@ -7,19 +9,24 @@
 @primary-tint-7: #0a1028;
 @primary-tint-8: #060a21;
 
+@success: #16825d;
+@success-tint-1: #199c6e;
+@danger: #d02029;
+@danger-tint-1: #e0212a;
+
+// Attribute / Property definitions
+
 @text-color: @primary-tint-1;
 @subtext-color: @primary-tint-2;
 
 @container-bg-color: @primary-tint-7;
 @container-fg-color: @primary-tint-6;
+@container-fg-color-bright: @primary-tint-5;
+
 @outline-color: @primary-tint-8;
 @shadow-color: black;
 @box-shadow: 0px 3px 4px @shadow-color;
+@border-radius: 5px;
 
 @scrollbar-bg-color: @container-fg-color;
 @scrollbar-thumb-color: @primary-tint-4;
-
-@success: #16825d;
-@success-tint-1: #199c6e;
-@danger: #d02029;
-@danger-tint-1: #e0212a;
