+++ src/components/browser/styles.less
@@ -33,7 +33,7 @@
 
 		.tab-item {
 			margin-right: 4px;
-			border-radius: 4px;
+			border-radius: @border-radius;
 			border-bottom-left-radius: 0;
 			border-bottom-right-radius: 0;
 			padding: 0px 8px;
