+++ src/components/explorer/styles.less
@@ -28,7 +28,7 @@
 .ContextMenu {
 	background-clip: padding-box;
 	background-color: #fff;
-	border-radius: 4px;
+	border-radius: @border-radius;
 	box-shadow: @box-shadow;
 	left: 0px;
 	list-style-type: none;
