--- src/components/panes/editor/ResultViewer/styles.less
@@ -1,137 +1,0 @@
-@import "../../../../common/styles/variables.less";
-
-.updates-list {
-	overflow: auto;
-	max-height: 250px;
-
-	.item {
-		margin-top: 8px;
-	}
-}
-
-.result-viewer {
-	display: flex;
-	flex-direction: column;
-	flex-grow: 1;
-	min-height: 0;
-
-	@result-viewer-header-height: 40px;
-
-	.result-viewer-header {
-		display: flex;
-		gap: 4px;
-		align-self: end;
-		align-items: center;
-		justify-content: flex-end;
-		flex-grow: 1;
-		min-height: @result-viewer-header-height;
-		padding: 2px;
-		width: 40%;
-
-		.header-item {
-			input {
-				width: 50px;
-				text-align: center;
-			}
-		}
-	}
-	.container {
-		display: flex;
-		min-height: 0;
-		width: 100%;
-		padding: 4px;
-	}
-
-	.tree-viewer {
-		display: flex;
-		flex-direction: column;
-		flex-grow: 1;
-		min-height: 0;
-
-		@tree-view-header-height: 30px;
-
-		.header {
-			min-height: @tree-view-header-height;
-			gap: 10px;
-		}
-
-		.content {
-			overflow-y: scroll;
-			overflow-x: hidden;
-			display: flex;
-			flex-direction: column;
-			flex-grow: 0;
-			height: calc(
-				100% - (@tree-view-header-height + @result-viewer-header-height)
-			);
-
-			.document-header-buttons {
-				display: flex;
-				* {
-					padding: 4px;
-				}
-			}
-		}
-
-		.context-menu {
-			display: flex;
-			flex-grow: 1;
-		}
-
-		.new-field-row {
-			display: flex;
-		}
-
-		.content-row {
-			@content-row-modified-width: 20px;
-
-			display: flex;
-			flex-grow: 1;
-
-			.button-container {
-				display: flex;
-			}
-
-			.left {
-				width: 50%;
-			}
-
-			.right {
-				width: 50%;
-			}
-
-			.modified {
-				display: flex;
-				justify-content: center;
-				align-items: center;
-				width: @content-row-modified-width;
-			}
-		}
-
-		.switchable-input {
-			display: flex;
-			flex-grow: 1;
-			align-items: center;
-
-			.object-id {
-				display: flex;
-			}
-
-			.switchable-input-child {
-				display: flex;
-				flex-grow: 1;
-			}
-		}
-	}
-
-	.json-viewer {
-		overflow-y: scroll;
-		overflow-x: hidden;
-		display: flex;
-		flex-direction: column;
-		width: 100%;
-		flex-grow: 0;
-		min-height: 0;
-		white-space: pre-wrap;
-	}
-}
