+++ src/components/panes/editor/result-viewer/ResultViewer.stories.tsx
