+++ src/components/panes/editor/result-viewer/styles.less
@@ -1,0 +1,249 @@
+@import "../../../../common/styles/variables.less";
+
+.updates-list {
+	overflow: auto;
+	max-height: 250px;
+
+	.item {
+		margin-top: 8px;
+	}
+}
+
+.result-viewer {
+	display: flex;
+	flex-direction: column;
+	flex-grow: 1;
+	min-height: 0;
+
+	@result-viewer-header-height: 40px;
+
+	.result-viewer-header {
+		display: flex;
+		gap: 4px;
+		align-self: end;
+		align-items: center;
+		justify-content: flex-end;
+		flex-grow: 1;
+		min-height: @result-viewer-header-height;
+		padding: 2px;
+		width: 40%;
+
+		.header-item {
+			input {
+				width: 50px;
+				text-align: center;
+			}
+		}
+	}
+
+	.container {
+		display: flex;
+		min-height: 0;
+		width: 100%;
+		padding: 4px;
+
+		.bp4-control-group {
+			align-items: center;
+		}
+
+		.bp4-input-group.bp4-small .bp4-input {
+			height: 21px;
+		}
+
+		.bp4-input {
+			height: 21px;
+		}
+	}
+
+	.tree-viewer {
+		display: flex;
+		flex-direction: column;
+		flex-grow: 1;
+		min-height: 0;
+		gap: 8px;
+
+		@tree-view-header-height: 32px;
+
+		.header {
+			min-height: @tree-view-header-height;
+
+			padding: 4px;
+			border-radius: @border-radius;
+			background-color: @container-bg-color;
+
+			/* To accomodate the overflow below the header (.content) */
+			margin-right: 10px;
+		}
+
+		.content {
+			overflow-y: scroll;
+			overflow-x: hidden;
+			display: flex;
+			flex-direction: column;
+			flex-grow: 0;
+			padding: 6px;
+			background-color: @container-bg-color;
+			border-radius: @border-radius;
+
+			height: calc(
+				100% - (@tree-view-header-height + @result-viewer-header-height + 8px)
+			);
+
+			.document-header-buttons {
+				display: flex;
+				* {
+					padding: 4px;
+				}
+			}
+		}
+
+		.context-menu {
+			display: flex;
+			flex-grow: 1;
+		}
+
+		.new-field-row {
+			display: flex;
+		}
+
+		.content-row {
+			@content-row-modified-width: 30px;
+
+			display: flex;
+			flex-grow: 1;
+			justify-content: start;
+			align-items: center;
+			padding-top: 2px;
+
+			min-height: 34px;
+
+			.button-container {
+				display: flex;
+			}
+
+			.left {
+				width: 50%;
+				height: 34px;
+				display: flex;
+				align-items: center;
+				padding: 4px 8px;
+				border-radius: @border-radius;
+				border: 1px solid @outline-color;
+				&:focus,
+				&:active,
+				&:hover {
+					background-color: @container-fg-color-bright;
+				}
+			}
+
+			.right {
+				width: 50%;
+				height: 34px;
+				display: flex;
+				align-items: center;
+				padding: 4px 8px;
+				border-radius: @border-radius;
+				border: 1px solid @outline-color;
+				&:focus,
+				&:active,
+				&:hover {
+					background-color: @container-fg-color-bright;
+				}
+			}
+
+			.modified {
+				display: flex;
+				justify-content: center;
+				height: 100%;
+				align-items: center;
+				width: @content-row-modified-width;
+			}
+		}
+
+		.switchable-input {
+			display: flex;
+			flex-grow: 1;
+			align-items: center;
+			gap: 8px;
+
+			span {
+				cursor: pointer;
+			}
+
+			.object-id {
+				display: flex;
+				align-items: center;
+			}
+
+			.switchable-input-child {
+				display: flex;
+				flex-grow: 1;
+				align-items: center;
+			}
+		}
+
+		.document-tree {
+			display: flex;
+			flex-direction: column;
+			width: 100%;
+			border-radius: @border-radius;
+			background-color: @container-bg-color;
+
+			.item {
+				cursor: pointer;
+				display: flex;
+				flex-direction: row;
+				align-items: center;
+				width: 100%;
+				height: 34px;
+				border-radius: @border-radius;
+				background-color: @container-fg-color;
+				border: 1px solid @outline-color;
+
+				.handle {
+					padding: 2px;
+					margin: 2px;
+				}
+
+				&.primary {
+					background-color: @container-fg-color;
+					padding: 4px;
+					height: 38px;
+					&:hover {
+						background-color: @container-fg-color-bright;
+					}
+				}
+
+				&:hover {
+					background-color: @container-fg-color-bright;
+				}
+
+				.heading {
+					width: 100%;
+					flex-grow: 1;
+					min-width: 0;
+				}
+
+				.top-right-element {
+					align-self: center;
+				}
+			}
+
+			.panel {
+				padding-left: 8px;
+				padding-bottom: 8px;
+			}
+		}
+	}
+
+	.json-viewer {
+		overflow-y: scroll;
+		overflow-x: hidden;
+		display: flex;
+		flex-direction: column;
+		width: 100%;
+		flex-grow: 0;
+		min-height: 0;
+		white-space: pre-wrap;
+	}
+}
