+++ src/index.less
@@ -3,7 +3,8 @@
 @import "./common/styles/variables.less";
 
 *::-webkit-scrollbar {
-	background-color: @container-fg-color;
+	background-color: @scrollbar-bg-color;
+	// border-radius: @border-radius;
 	width: 10px;
 	height: 10px;
 }
@@ -14,7 +15,7 @@
 
 *:hover::-webkit-scrollbar-thumb {
 	background-color: @scrollbar-thumb-color;
-	border-radius: 6px;
+	border-radius: @border-radius;
 	border: 3px solid @scrollbar-bg-color;
 	opacity: 100%;
 	transition: opacity 3s;
